requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=openblas
bottleneck=1.3.5=py39heec5a64_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
ca-certificates=2023.05.30=hca03da5_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py39h525c30c_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py39h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
giflib=5.2.1=h80987f9_3
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.15.0=pyh31c8845_0
jedi=0.19.0=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
jpeg=9e=h80987f9_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.3.0=py39hca03da5_0
kiwisolver=1.4.4=py39h313beb8_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libsodium=1.0.18=h27ca646_1
libtiff=4.5.1=h313beb8_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
matplotlib=3.7.2=py39hca03da5_0
matplotlib-base=3.7.2=py39h46d7db6_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numexpr=2.8.4=py39h79ee842_1
numpy=1.25.2=py39h3b2db8e_0
numpy-base=1.25.2=py39ha9811e2_0
openssl=3.0.10=h1a28f6b_2
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py39h46d7db6_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py39h313beb8_0
pip=23.2.1=py39hca03da5_0
platformdirs=3.10.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.0=py39h1a28f6b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.0.9=py39hca03da5_0
python=3.9.17=hb885b13_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2022.7=py39hca03da5_0
pyzmq=25.1.0=py39h313beb8_0
readline=8.2=h1a28f6b_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
seaborn=0.12.2=py39hca03da5_0
setuptools=68.0.0=py39hca03da5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tornado=6.1=py39h5161555_1
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py39hca03da5_0
xz=5.4.2=h80987f9_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.11.0=py39hca03da5_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
